refactor: KIS 실시간 주가 데이터 파싱 변경

+++ src/main/java/com/stockleague/backend/stock/dto/response/stock/StockPriceDto.java
@@ -30,7 +30,7 @@ public record StockPriceDto(
         int priceChange,
 
         @Schema(description = "전일 대비율", example = "2.15")
-        int pricePercent,
+        double pricePercent,
 
         @Schema(description = "등락 부호 (1:상승, 2:하락, 3:보합)", example = "1")
         int changeSign,
